+++ crates/objc2/CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added `IsMainThreadOnly::mtm`.
 * Added `DowncastTarget`, `AnyObject::downcast_ref` and `Retained::downcast`
   to allow safely casting between Objective-C objects.
+* Implemented more `fmt` traits on `Retained`.
+* Implemented `Extend` trait on `Retained`.
 
 ### Changed
 * **BREAKING**: Changed how you specify a class to only be available on the